requirements.txt
# Core Dependencies
playwright>=1.40.0
playwright-stealth>=1.0.6  # Optional - enhances stealth capabilities

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP & Networking
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pydantic>=2.5.0
jsonschema>=4.20.0
pandas>=2.1.0

# CLI & UI
rich>=13.7.0
typer>=0.9.0
streamlit>=1.29.0
plotly>=5.18.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2

# Utilities
faker>=21.0.0
tqdm>=4.66.0
python-dateutil>=2.8.2

# CAPTCHA solving (optional)
# 2captcha-python>=1.2.0

# Proxy support (optional)
# requests[socks]>=2.31.0
